Authentication/requirements.txt
django-cors-headers
djangorestframework
djangorestframework_simplejwt
Django